Adjust bulk photo nodes form markup
The markup generated by forms api makes more sense if there are two distinct
fieldset. One for the photo node information, and one for the bulk settings.
I've refactored the form elements so that they are both wrapped in fieldsets,
which will be floated left and right, respectively. I've also added an
"attributes" element, to give the fieldsets a reasonable class to key off of.
These classes will probably change, but that shouldn't affect anything outside
of the theme layer.